Move logic for static dir CopyWebpackPlugin to appropriate file
