+++ packages/docusaurus/src/webpack/server.ts
@@ -6,9 +6,11 @@
  */
 
 import path from 'path';
+import fs from 'fs-extra';
 import merge from 'webpack-merge';
 import {NODE_MAJOR_VERSION, NODE_MINOR_VERSION} from '@docusaurus/utils';
 import WebpackBar from 'webpackbar';
+import CopyWebpackPlugin from 'copy-webpack-plugin';
 import {createBaseConfig} from './base';
 import type {Props} from '@docusaurus/types';
 import type {Configuration} from 'webpack';
@@ -36,6 +38,46 @@ export default async function createServerConfig(params: {
         name: 'Server',
         color: 'yellow',
       }),
-    ],
+      await createStaticDirectoriesCopyPlugin(params),
+    ].filter(Boolean),
+  });
+}
+
+async function createStaticDirectoriesCopyPlugin({props}: {props: Props}) {
+  const {
+    outDir,
+    siteDir,
+    siteConfig: {staticDirectories: staticDirectoriesOption},
+  } = props;
+
+  // The staticDirectories option can contain empty directories, or non-existent
+  // directories (e.g. user deleted `static`). Instead of issuing an error, we
+  // just silently filter them out, because user could have never configured it
+  // in the first place (the default option should always "work").
+  const staticDirectories: string[] = (
+    await Promise.all(
+      staticDirectoriesOption.map(async (dir) => {
+        const staticDir = path.resolve(siteDir, dir);
+        if (
+          (await fs.pathExists(staticDir)) &&
+          (await fs.readdir(staticDir)).length > 0
+        ) {
+          return staticDir;
+        }
+        return '';
+      }),
+    )
+  ).filter(Boolean);
+
+  if (staticDirectories.length === 0) {
+    return undefined;
+  }
+
+  return new CopyWebpackPlugin({
+    patterns: staticDirectories.map((dir) => ({
+      from: dir,
+      to: outDir,
+      toType: 'dir',
+    })),
   });
 }
